feat: ShareDiary 프로토콜 추가 및 delegate 삭제, 오류수정
