+++ Diary/Protocol/AlertDisplayble.swift
@@ -2,7 +2,7 @@
 //  AlertDisplayble.swift
 //  Diary
 //
-//  Created by 1 on 2023/09/05.
+//  Created by Max, Hemg on 2023/09/05.
 //
 
 import UIKit
